+++ git-pack/src/cache/delta/iter.rs
@@ -1,4 +1,4 @@
-use crate::tree::{Item, Tree};
+use crate::cache::delta::{Item, Tree};
 
 /// Iteration
 impl<T> Tree<T> {