+++ git-pack/CHANGELOG.md
@@ -1,6 +1,6 @@
 ### 0.9.0 (2021-08-??)
 
-- **renames / moves**
+- **renames / moves / visibility**
    - `find::Find`  and `find::FindExt` only in `Find` and `FindExt` (not in `find` anymore)
    - `data::output::count::Count` -> `data::output::Count`
    - `data::output::entry::Entry` -> `data::output::Entry`
@@ -9,6 +9,7 @@
    - `Find::find()-> `Find::try_find()`
    - `bundle::Bundle` -> `Bundle`
    - `bundle::Error` -> `bundle::init::Error`
+   - `pub tree::` -> `pub(crate) cache::delta::`
  
 * **new methods**
    - `Find::find_tag_iter()`